Use linked version of libheif-js instead
