+++ packages/design-system/src/typings/styled.d.ts
@@ -25,5 +25,6 @@ import 'styled-components';
 import type { Theme } from '../theme';
 
 declare module 'styled-components' {
+  // eslint-disable-next-line @typescript-eslint/no-empty-interface -- WAI.
   export interface DefaultTheme extends Theme {}
 }
